[CI] Update dockerfiles that build-wasm to include bitcoinsuite-chronik-client
Summary:
See latest Cashtab CI failure

With a new module in the monorepo that is part of the top level cargo.toml scope, all CI apps that need this scope are broken unless they include everything in this workspace.

Since this has happened twice now in less than a month ... we should have some better way of catching this. However, the issue is known and the fix is simple enough. Getting CI to build all dockerfiles every time cargo.toml changes, mb?

Adding new stuff to the top level cargo.toml is not expected to be this common in the future, so the bigger risk is this doesn't happen again for 18 months and then when it does, we don't remember what's going on.

Test Plan: `grep "FROM rust:1.76.0 AS wasmbuilder" .` and confirm all dockerfiles that build from rust have this added in the same way it was added for the explorer

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17384