[Cashtab] Improve portability of getFormattedFiatPrice
Summary:
This function takes `settings` as a param because all of it's callsites happend to have it available. However it is only using the fiatCurrency key of settings to know what ticker to print.

Modify so we only take this ticker as as param and update callsights.

This is part of a broader rethink to simplify searching and sorting agora offers by price.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D17385