[Cashtab] No more utils folder
Summary:
`cashMethods` was a catch-all file for misc functions used for ... all kinds of thing over the history of Cashtab.

It quickly became too generic. Then many of the functions were solving problems that we shouldn't have been solving at this level of the stack.

Now, these problems are finally all handled in libraries or the appropriate part of the app. So, get rid of cashMethods.js which is no longer used. Rename `utils` to `formatting` since now all we handle here is formatting. Update imports.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D17416