[Cashtab] Improve notification styles
Summary:
We cannot view links with the dark mode notifications on the old settings. Fix the css.

In solving this problem, noticed we really do not need special styled components for links. Get rid of them. Handle all notifications styles globally.

Note this is something we can more easily do now with the latest version of `react-toastify`. While this is a "bug fix" in some ways... I think it's okay that Cashtab does not test the css colors of all components.

Test Plan:
`npm test`, see screenshots

Current:
{F12701416}

after this diff:
{F12701417}

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17481