[mock-chronik-client] Update README
Summary:
The README is totally wrong and obsolete. With ts, everything is defined in the class.

Types are stable but the module is still not feature complete. I do not think it's worth updating the README yet as the app is being built and only serious devs use this app, who would get better info from the tests.

Test Plan: proofread

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D17506