[Cashtab] Make agora screen more polite
Summary: Tweak the politeness params on agora so it takes longer to load everything but the screen is more usable, less likely to crash

Test Plan:
Can `npm test` but really, just observe we are asking for a lot less and less often.

First implementation of these params was trying to keep the same loading behavior. Now that loading behavior is causing the app to fail most every time. So, slow things down.

We may tweak a couple times to get it right but this is also still an interim thing. We need proper lazy loading and search to solve the problem.

Reviewers: #bitcoin_abc, johnkuney

Reviewed By: #bitcoin_abc, johnkuney

Differential Revision: https://reviews.bitcoinabc.org/D17511