[Cashtab] Prevent creation of unacceptable agora offers
Summary: I have validation for this that "should" work but clearly is not. Ultimately the validation belongs in `ecash-agora`. Because this may take time to debug properly, for now make sure we do not allow this type of offer to be created in Cashtab.

Test Plan:
`npm test`

I have not been able to repeat the issue due to the validation. But the issue is clearly happening because users have reported it, and say they are on the latest version.

So, I am not able to test it specifically in the tests because I am not sure how to outsmart the validation in the tests. I think it involves changing wallets, or switching tokens, or scrolling the mouse while the modal is up. Still looking into it.

`npm test` confirms we do not break existing functionality with this diff.

Check the test site and try to create an unacceptable offer at https://cashtab-local-dev.netlify.app/

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D17515